+++ src/core/jit/exec.rs
@@ -5,21 +5,26 @@ use std::sync::{Arc, Mutex};
 use derive_getters::Getters;
 use futures_util::future::join_all;
 
+use super::context::Context;
+use super::synth::Synthesizer;
 use super::{Children, ExecutionPlan, Field, Request, Store};
 use crate::core::ir::model::IR;
 use crate::core::json::JsonLike;
 
-pub struct QueryExecutor<Synth, Exec> {
+///
+/// Default GraphQL executor that takes in a GraphQL Request and produces a
+/// GraphQL Response
+pub struct Executor<Synth, IRExec> {
     plan: ExecutionPlan,
     synth: Synth,
-    exec: Exec,
+    exec: IRExec,
 }
 
-impl<Input, Output, Error, Synth, Exec> QueryExecutor<Synth, Exec>
+impl<Input, Output, Error, Synth, Exec> Executor<Synth, Exec>
 where
     Output: JsonLike<Output = Output>,
     Synth: Synthesizer<Value = Result<Output, Error>>,
-    Exec: Executor<Input = Input, Output = Output, Error = Error>,
+    Exec: IRExecutor<Input = Input, Output = Output, Error = Error>,
 {
     pub fn new(plan: ExecutionPlan, synth: Synth, exec: Exec) -> Self {
         Self { plan, synth, exec }
@@ -28,8 +33,7 @@ where
     async fn execute_inner(&self, request: Request<Input>) -> Store<Result<Output, Error>> {
         let store: Arc<Mutex<Store<Result<Output, Error>>>> =
             Arc::new(Mutex::new(Store::new(self.plan.size())));
-        let mut ctx =
-            QueryExecutorInner::new(request, store.clone(), self.plan.to_owned(), &self.exec);
+        let mut ctx = ExecutorInner::new(request, store.clone(), self.plan.to_owned(), &self.exec);
         ctx.execute().await;
 
         let store = mem::replace(&mut *store.lock().unwrap(), Store::new(0));
@@ -43,17 +47,17 @@ where
 }
 
 #[derive(Getters)]
-pub struct QueryExecutorInner<'a, Input, Output, Error, Exec> {
+struct ExecutorInner<'a, Input, Output, Error, Exec> {
     request: Request<Input>,
     store: Arc<Mutex<Store<Result<Output, Error>>>>,
     plan: ExecutionPlan,
     exec: &'a Exec,
 }
 
-impl<'a, Input, Output, Error, Exec> QueryExecutorInner<'a, Input, Output, Error, Exec>
+impl<'a, Input, Output, Error, Exec> ExecutorInner<'a, Input, Output, Error, Exec>
 where
     Output: JsonLike<Output = Output>,
-    Exec: Executor<Input = Input, Output = Output, Error = Error>,
+    Exec: IRExecutor<Input = Input, Output = Output, Error = Error>,
 {
     fn new(
         request: Request<Input>,
@@ -66,7 +70,7 @@ where
 
     async fn execute(&mut self) {
         join_all(self.plan.as_children().iter().map(|field| async {
-            let ctx = EvaluationContext::new(&self.request);
+            let ctx = Context::new(&self.request);
             self.execute_field(field, &ctx, false).await
         }))
         .await;
@@ -75,7 +79,7 @@ where
     async fn execute_field<'b>(
         &'b self,
         field: &'b Field<Children>,
-        ctx: &'b EvaluationContext<'b, Input, Output>,
+        ctx: &'b Context<'b, Input, Output>,
         is_multi: bool,
     ) -> Result<(), Error> {
         if let Some(ir) = &field.ir {
@@ -121,58 +125,15 @@ where
     }
 }
 
-#[derive(Getters)]
-pub struct EvaluationContext<'a, Input, Output> {
-    request: &'a Request<Input>,
-    parent: Option<&'a Output>,
-    value: Option<&'a Output>,
-}
-
-impl<'a, Input, Output> Clone for EvaluationContext<'a, Input, Output> {
-    fn clone(&self) -> Self {
-        Self {
-            request: self.request,
-            parent: self.parent,
-            value: self.value,
-        }
-    }
-}
-
-impl<'a, Input, Output> EvaluationContext<'a, Input, Output> {
-    fn new(request: &'a Request<Input>) -> Self {
-        Self { request, parent: None, value: None }
-    }
-
-    fn with_parent_value(&self, value: &'a Output) -> Self {
-        Self {
-            request: self.request,
-            parent: self.parent,
-            value: Some(value),
-        }
-    }
-
-    fn with_value(&self, value: &'a Output) -> Self {
-        Self {
-            request: self.request,
-            parent: self.parent,
-            value: Some(value),
-        }
-    }
-}
-
-pub trait Synthesizer {
-    type Value;
-    fn synthesize(&self, store: &Store<Self::Value>) -> Self::Value;
-}
-
+/// Executor for IR
 #[async_trait::async_trait]
-pub trait Executor {
+pub trait IRExecutor {
     type Input;
     type Output;
     type Error;
     async fn execute<'a>(
         &'a self,
         ir: &'a IR,
-        ctx: &'a EvaluationContext<'a, Self::Input, Self::Output>,
+        ctx: &'a Context<'a, Self::Input, Self::Output>,
     ) -> Result<Self::Output, Self::Error>;
 }
